[#977] - Actualiza ubicación de mocks para respuestas de API de Clarity
